Refactor: Improve code quality
