+++ app/Http/Controllers/Api/DashboardController.php
@@ -2,7 +2,6 @@
 
 namespace App\Http\Controllers\Api;
 
-use App\Http\Controllers\Controller;
 use App\Models\ChatRoom;
 use App\Models\Message;
 use App\Models\PublicChat;
@@ -10,7 +9,6 @@ use Illuminate\Http\Request;
 
 class DashboardController extends ApiController
 {
-
     public function getDashboard(Request $request)
     {
         $user = $request->user();
@@ -25,7 +23,7 @@ class DashboardController extends ApiController
             'userProfile' => $userProfile,
             'messageCount' => $messageCount,
             'chatRoomCount' => $chatRoomCount,
-            'publicChats' => $publicChats
+            'publicChats' => $publicChats,
         ];
 
         return $this->respond($response);
